create: VWBLBase to extends both erc class
