+++ src/vwbl/blockchain/ERC1155/VWBLProtocol.ts
@@ -5,7 +5,7 @@ import { AbiItem } from "web3-utils";
 import vwbl1155 from "../../../contract/VWBLERC1155.json";
 import vwbl1155IPFS from "../../../contract/VWBLERC1155SupportIPFS.json";
 
-export class VWBLNFT {
+export class VWBLERC1155NFT {
   private contract: Contract;
   private web3: Web3;
