+++ src/vwbl/index.ts
@@ -1,4 +1,5 @@
-export * from "./VWBL";
+export * from "./erc721/VWBL";
+export * from "./erc1155/VWBL";
 export * from "./metadata";
 export * from "./types";
 export * from "./blockchain";
